add new bundles, move libs from c.g.c.t.e.login to the new bundle
