+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.ui/META-INF/MANIFEST.MF
@@ -17,6 +17,7 @@ Import-Package: com.google.api.client.auth.oauth2,
  com.google.cloud.tools.eclipse.appengine.deploy.standard,
  com.google.cloud.tools.eclipse.appengine.facets,
  com.google.cloud.tools.eclipse.appengine.ui,
+ com.google.cloud.tools.eclipse.googleapis,
  com.google.cloud.tools.eclipse.login,
  com.google.cloud.tools.eclipse.login.ui,
  com.google.cloud.tools.eclipse.projectselector,
