+++ plugins/com.google.cloud.tools.eclipse.googleapis.test/.classpath
