+++ plugins/com.google.cloud.tools.eclipse.googleapis.test/META-INF/MANIFEST.MF
@@ -1,0 +1,13 @@
+Manifest-Version: 1.0
+Bundle-ClassPath: .
+Bundle-ManifestVersion: 2
+Bundle-Name: Tests for Cloud Tools for Eclipse Google API Client Libraries
+Bundle-RequiredExecutionEnvironment: JavaSE-1.7
+Bundle-SymbolicName: com.google.cloud.tools.eclipse.googleapis.test
+Bundle-Vendor: Google Inc.
+Bundle-Version: 0.1.0.qualifier
+Fragment-Host: com.google.cloud.tools.eclipse.googleapis
+Require-Bundle: org.junit;bundle-version="4.12.0"
+Import-Package: org.mockito;provider=google;version="1.10.19",
+ org.mockito.runners;provider=google;version="1.10.19",
+ org.mockito.stubbing;provider=google;version="1.10.19"
