+++ plugins/com.google.cloud.tools.eclipse.googleapis.test/build.properties
@@ -1,7 +1,6 @@
 source.. = src/
 output.. = target/classes/
 bin.includes = META-INF/,\
-               .,\
-               plugin.properties
+               .
 javacSource=1.7
-javacTarget=1.7               
+javacTarget=1.7
