+++ plugins/com.google.cloud.tools.eclipse.googleapis/META-INF/MANIFEST.MF
@@ -1,0 +1,42 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Cloud Tools for Eclipse Google API Client Libraries
+Bundle-SymbolicName: com.google.cloud.tools.eclipse.googleapis;singleton:=true
+Bundle-Version: 0.1.0.qualifier
+Bundle-Vendor: Google Inc.
+Bundle-Localization: plugin
+Bundle-RequiredExecutionEnvironment: JavaSE-1.7
+Bundle-ActivationPolicy: lazy
+Bundle-Classpath: .,
+ lib/google-api-client-1.22.0.jar,
+ lib/google-http-client-1.22.0.jar,
+ lib/google-http-client-jackson-1.22.0.jar,
+ lib/google-api-services-appengine-v1-rev9-1.21.0.jar,
+ lib/google-api-services-cloudresourcemanager-v1-rev355-1.22.0.jar,
+ lib/google-api-services-oauth2-v2-rev114-1.22.0.jar,
+ lib/google-oauth-client-1.22.0.jar,
+ lib/google-oauth-client-java6-1.22.0.jar,
+ lib/google-oauth-client-jetty-1.22.0.jar,
+ lib/plugins-login-common-0.1.1-SNAPSHOT.jar
+Require-Bundle: com.google.guava;bundle-version="[20.0.0,21.0.0)"
+Export-Package: com.google.api.client.auth.oauth2,
+ com.google.api.client.extensions.jetty.auth.oauth2,
+ com.google.api.client.googleapis,
+ com.google.api.client.googleapis.auth.oauth2,
+ com.google.api.client.googleapis.json,
+ com.google.api.client.googleapis.services,
+ com.google.api.client.googleapis.services.json,
+ com.google.api.client.googleapis.testing.json,
+ com.google.api.client.http,
+ com.google.api.client.http.javanet,
+ com.google.api.client.json,
+ com.google.api.client.json.jackson,
+ com.google.api.client.util,
+ com.google.api.services.appengine.v1,
+ com.google.api.services.appengine.v1.model,
+ com.google.api.services.cloudresourcemanager,
+ com.google.api.services.cloudresourcemanager.model,
+ com.google.cloud.tools.eclipse.googleapis,
+ com.google.cloud.tools.login;version="0.1.1"
+Import-Package: com.google.cloud.tools.eclipse.util,
+ org.codehaus.jackson
