+++ plugins/com.google.cloud.tools.eclipse.googleapis/build.properties
@@ -1,0 +1,15 @@
+source.. = src/
+output.. = target/classes/
+bin.includes = META-INF/,\
+               .,\
+               plugin.properties,\
+               lib/,\
+               lib/google-api-services-appengine-v1-rev9-1.21.0.jar,\
+               lib/google-api-services-cloudresourcemanager-v1-rev355-1.22.0.jar,\
+               lib/google-api-services-oauth2-v2-rev114-1.22.0.jar,\
+               lib/google-oauth-client-1.22.0.jar,\
+               lib/google-oauth-client-java6-1.22.0.jar,\
+               lib/google-oauth-client-jetty-1.22.0.jar,\
+               lib/plugins-login-common-0.1.1-SNAPSHOT.jar
+javacSource=1.7
+javacTarget=1.7
