+++ plugins/com.google.cloud.tools.eclipse.googleapis/lib/google-http-client-1.22.0.jar
