+++ plugins/com.google.cloud.tools.eclipse.googleapis/lib/google-http-client-jackson-1.22.0.jar
