+++ plugins/com.google.cloud.tools.eclipse.googleapis/lib/google-oauth-client-java6-1.22.0.jar
