+++ plugins/com.google.cloud.tools.eclipse.googleapis/lib/google-oauth-client-jetty-1.22.0.jar
