+++ plugins/com.google.cloud.tools.eclipse.googleapis/plugin.properties
@@ -1,0 +1,2 @@
+pluginName=Cloud Tools for Eclipse Google API Client Libraries
+providerName=Google Inc.
