+++ plugins/com.google.cloud.tools.eclipse.login.test/META-INF/MANIFEST.MF
@@ -8,7 +8,9 @@ Bundle-Vendor: Google Inc.
 Bundle-Version: 0.1.0.qualifier
 Fragment-Host: com.google.cloud.tools.eclipse.login
 Require-Bundle: org.junit;bundle-version="4.12.0"
-Import-Package: com.google.cloud.tools.eclipse.test.util,
+Import-Package: com.google.api.client.http.javanet,
+ com.google.api.client.json.jackson,
+ com.google.cloud.tools.eclipse.test.util,
  com.google.cloud.tools.eclipse.test.util.ui,
  org.eclipse.swtbot.swt.finder,
  org.eclipse.swtbot.swt.finder.exceptions,
