+++ plugins/com.google.cloud.tools.eclipse.login/src/com/google/cloud/tools/eclipse/login/ui/AccountSelector.java
@@ -17,10 +17,10 @@
 package com.google.cloud.tools.eclipse.login.ui;
 
 import com.google.api.client.auth.oauth2.Credential;
-import com.google.api.client.util.Strings;
 import com.google.cloud.tools.eclipse.login.IGoogleLoginService;
 import com.google.cloud.tools.login.Account;
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Strings;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
