+++ plugins/com.google.cloud.tools.eclipse.projectselector/META-INF/MANIFEST.MF
@@ -1,7 +1,5 @@
 Manifest-Version: 1.0
-Bundle-ClassPath: .,
- lib/google-api-services-cloudresourcemanager-v1-rev355-1.22.0.jar,
- lib/google-api-services-appengine-v1-rev9-1.21.0.jar
+Bundle-ClassPath: .
 Bundle-ManifestVersion: 2
 Bundle-Name: %pluginName
 Bundle-SymbolicName: com.google.cloud.tools.eclipse.projectselector;singleton:=true
@@ -22,6 +20,11 @@ Import-Package: com.google.api.client.auth.oauth2,
  com.google.api.client.json,
  com.google.api.client.json.jackson,
  com.google.api.client.util,
+ com.google.api.services.appengine.v1,
+ com.google.api.services.appengine.v1.model,
+ com.google.api.services.cloudresourcemanager,
+ com.google.api.services.cloudresourcemanager.model,
+ com.google.cloud.tools.eclipse.googleapis,
  com.google.cloud.tools.eclipse.login.ui,
  com.google.cloud.tools.eclipse.ui.util,
  com.google.cloud.tools.eclipse.ui.util.event,
