+++ pom.xml
@@ -58,6 +58,8 @@
     <module>plugins/com.google.cloud.tools.eclipse.appengine.validation.test</module>
     <module>plugins/com.google.cloud.tools.eclipse.appengine.ui</module>
     <module>plugins/com.google.cloud.tools.eclipse.appengine.ui.test</module>
+    <module>plugins/com.google.cloud.tools.eclipse.googleapis</module>
+    <module>plugins/com.google.cloud.tools.eclipse.googleapis.test</module>
     <module>plugins/com.google.cloud.tools.eclipse.preferences</module>
     <module>plugins/com.google.cloud.tools.eclipse.preferences.test</module>
     <module>plugins/com.google.cloud.tools.eclipse.projectselector</module>
